+++ src/components/Inputs/Checkbox/Checkbox.tsx
@@ -20,6 +20,9 @@ const Checkbox = ({
     labelWidth = 30,
     size = InputSize.MEDIUM,
     disabled = false,
+    readOnly = false,
+    noBottomMargin = false,
+    classNamesObj,
     ...otherProps
 }: CheckboxProps) => {
     const checkboxContainerRef = useRef<HTMLDivElement>(null);
@@ -30,6 +33,11 @@ const Checkbox = ({
     }, [checked]);
 
     const handleChange = (e: ChangeEvent<HTMLInputElement>): void => {
+        if (readOnly) {
+            e.preventDefault();
+            return;
+        }
+
         const _checked = e.target.checked;
 
         onChange && onChange(e, _checked);
@@ -37,10 +45,18 @@ const Checkbox = ({
     };
 
     return (
-        <InputContainer disabled={disabled} ref={checkboxContainerRef} className="m-checkbox-container" size={size} error={error}>
+        <InputContainer
+            disabled={disabled}
+            ref={checkboxContainerRef}
+            className={classNames("m-checkbox-container", classNamesObj?.container)}
+            size={size}
+            error={error}
+            noBottomMargin={noBottomMargin}
+        >
             <div style={getInputStyle(labelType as SimpleInputLabel, label, labelWidth, undefined)}>
-                <label className={classNames("m-checkbox-input-wrapper", { checked: isChecked })}>
+                <label className={classNames("m-checkbox-input-wrapper", classNamesObj?.inputWrapper, { checked: isChecked })}>
                     <input
+                        readOnly
                         className="m-checkbox-input"
                         type="checkbox"
                         checked={isChecked}
@@ -49,16 +65,23 @@ const Checkbox = ({
                         disabled={disabled}
                         {...otherProps}
                     />
-                    <span className={classNames("m-input", "m-checkbox", labelType)}>
+                    <span className={classNames("m-input", "m-checkbox", classNamesObj?.input, labelType)}>
                         <FontAwesomeIcon className="m-checkbox-check-icon" icon="check" />
                     </span>
                 </label>
             </div>
-            {label && <InputsLabel label={label} labelType={labelType} className="m-input-label" labelWidth={labelWidth} />}
+            {label && (
+                <InputsLabel
+                    label={label}
+                    labelType={labelType}
+                    className={classNames("m-input-label", classNamesObj?.label)}
+                    labelWidth={labelWidth}
+                />
+            )}
             {error && checkboxContainerRef.current && (
                 <InputError
                     style={getCheckboxErrorStyle(checkboxContainerRef.current, labelType as SimpleInputLabel, labelWidth)}
-                    className="checkbox"
+                    className={classNames("checkbox", classNamesObj?.error)}
                     error={error}
                 />
             )}