+++ src/components/Inputs/Dropdown/types.ts
@@ -2,6 +2,15 @@ import { FocusEvent } from "react";
 
 import * as GlobalInterfaces from "../../global-types";
 
+type DropdownClassNames = {
+    container?: string;
+    control?: string;
+    label?: string;
+    error?: string;
+    dropdownIndicatorIcon?: string;
+    clearIcon?: string;
+} & DropdownOptionsClassnames;
+
 type DropdownBaseProps<T> = {
     value?: DropdownValue<T>;
     name?: string;
@@ -19,6 +28,8 @@ type DropdownBaseProps<T> = {
     size?: `${GlobalInterfaces.InputSize}`;
     label?: string;
     labelType?: `${GlobalInterfaces.InputLabel}`;
+    noBottomMargin?: boolean;
+    classNamesObj?: DropdownClassNames;
 
     options?: T[];
 };
@@ -51,6 +62,12 @@ type DropdownChangeEventTarget<T> = EventTarget & {
     type: "dropdown";
 };
 
+type DropdownOptionsClassnames = {
+    dropdownOptions?: string;
+    dropdownOption?: string;
+    emptyDropdownOption?: string;
+};
+
 export type DropdownOptionsProps<T> = {
     filterElement: HTMLInputElement;
     uniqueDropdownId: string;
@@ -59,4 +76,5 @@ export type DropdownOptionsProps<T> = {
     value: T | undefined;
     valueKey: keyof T;
     labelKey: keyof T;
+    classNamesObj?: DropdownOptionsClassnames;
 };