+++ src/components/Inputs/Slider/types.ts
@@ -2,6 +2,14 @@ import * as GlobalInterfaces from "../../global-types";
 
 export type ValuePreviewType = "top-dynamic" | "bottom-dynamic" | "top-static" | "bottom-static";
 
+type SliderClassNames = {
+    container?: string;
+    input?: string;
+    label?: string;
+    error?: string;
+    valuePreview?: string;
+};
+
 export type SliderProps = {
     value?: number;
     min: number;
@@ -17,5 +25,8 @@ export type SliderProps = {
     labelType?: `${GlobalInterfaces.InputLabel}`;
     hideValuePreview?: boolean;
     disabled?: boolean;
+    readOnly?: boolean;
     valuePreviewType?: ValuePreviewType;
+    noBottomMargin?: boolean;
+    classNamesObj?: SliderClassNames;
 };