+++ src/components/Inputs/Textarea/types.ts
@@ -2,6 +2,13 @@ import { ChangeEvent, FocusEvent } from "react";
 
 import * as GlobalInterfaces from "../../global-types";
 
+type TextAreaClassNames = {
+    container?: string;
+    input?: string;
+    label?: string;
+    error?: string;
+};
+
 export type TextareaProps = {
     value?: string;
     name?: string;
@@ -12,10 +19,13 @@ export type TextareaProps = {
     defaultInternalValue?: string;
     autoFocus?: boolean;
     disabled?: boolean;
+    readOnly?: boolean;
     labelWidth?: GlobalInterfaces.LabelPercentageWidth;
     floatingInputWidth?: GlobalInterfaces.FloatingInputWidth;
     row?: number;
     size?: `${GlobalInterfaces.InputSize}`;
     label?: string;
     labelType?: `${GlobalInterfaces.InputLabel}`;
+    noBottomMargin?: boolean;
+    classNamesObj?: TextAreaClassNames;
 };