+++ CHANGELOG.md
@@ -1,14 +1,27 @@
+### 18.10.2024 - Release 1.0.4
+
+-   [[#37]](https://github.com/Marcin-Migdal/m-component-library/issues/37) **[PATCH]** Alert close only when overlay click started on overlay
+-   [[#36]](https://github.com/Marcin-Migdal/m-component-library/issues/36) **[PATCH]** Add hslToHex helper function
+-   [[#34]](https://github.com/Marcin-Migdal/m-component-library/issues/34) **[PATCH]** Fix clear icon in dropdown wrong color
+-   [[#35]](https://github.com/Marcin-Migdal/m-component-library/issues/35) **[PATCH]** Add noBottomMargin props to all inputs that on default have bottom margin
+-   [[#33]](https://github.com/Marcin-Migdal/m-component-library/issues/33) **[PATCH]** Add readOnly prop to inputs
+-   [[#55]](https://github.com/Marcin-Migdal/m-component-library/issues/55) **[PATCH]** Add classNamesObj to input components
+
+### 28.09-2024 - Release 1.0.3
+
+-   **[PATCH]** Fix colors
+
 ### 28.09-2024 - Release 1.0.2
 
--   Fix colors
+-   **[PATCH]** Fix colors
 
 ### 28.09-2024 - Release 1.0.2
 
--   Fix colors
+-   **[PATCH]** Fix colors
 
 ### 28.09-2024 - Release 1.0.1
 
--   Add more color variables
+-   **[PATCH]** Add more color variables
 
 ### 28.09-2024 - Release 1.0.0